+++ app/templates/api/services/AuthService.js
@@ -1,0 +1,3 @@
+var AuthFactory = require('./auth/AuthFactory');
+
+module.exports = new AuthFactory().create('bcrypt');